refactor(no-code): extract TaskDrawer and PluginPicker into standalone components
- TaskDrawer.vue: self-contained three-column drawer (Inputs / Properties /
  Outputs). Injects FULL_SOURCE_INJECTION_KEY to derive flow inputs and reads
  pluginsStore for schema-driven output list. Receives a creatingTask prop and
  emits close.

- PluginPicker.vue: self-contained plugin browser with search, kind filter
  (All / Tasks / Triggers), collapsible groups with show-more/fewer. Exports
  PluginEntry type and PLUGIN_DRAG_TYPE constant for cross-component drag data.
  Rows are draggable (HTML5 drag-and-drop, dataTransfer keyed by PLUGIN_DRAG_TYPE)
  and emit addPlugin on click.

- NoCodeWorkspace.vue: reduced to orchestration. Flow sections that accept
  drops (tasks, triggers, errors, finally, afterExecution, pluginDefaults) are
  wrapped in droppable-section divs. dragover highlights the target section;
  drop deserialises the PluginEntry from dataTransfer and calls addPlugin with
  the explicit target section, overriding kind-based defaulting.

Adds no_code.workspace.drop_hint i18n key.
